#10460368 Реализовал логику добавления заказа в кэш. Тест пока не писал.

+++ LeokaEstetica.Platform.Models/Dto/Common/Cache/CreateOrderCache.cs
@@ -31,5 +31,5 @@ public class CreateOrderCache
     /// Месяц.
     /// </summary>
     [ProtoMember(4)]
-    public int Month { get; set; }
+    public short Month { get; set; }
 }
